--- cset-workflow/app/db/file/redis.conf
@@ -1,17 +1,0 @@
-# Redis configuration file.
-#
-# Note that in order to read the configuration file, Redis must be
-# started with the file path as first argument:
-#
-# ./redis-server /path/to/redis.conf
-
-# Bind to all interfaces on port 6379.
-bind * -::*
-port 6379
-
-# Disable security as we are on a closed network.
-protected-mode no
-
-# Save dumps to dump.rdb in the current working directory.
-dbfilename dump.rdb
-dir ./
