--- cset-workflow/app/db/rose-app.conf
@@ -1,2 +1,0 @@
-[command]
-default=app_env_wrapper run-redis-server.py
