+++ cset-workflow/app/fetch_fcst/rose-app.conf
@@ -1,0 +1,2 @@
+[command]
+default=echo "Please set FETCH_FCST_OPT_CONF to your storage system."; false
