+++ cset-workflow/app/metplus_ascii2nc/file/ASCII2NC_python_embedding.conf
@@ -1,0 +1,101 @@
+[dir]
+INPUT_BASE = {ENV[METPLUS_OBS_DIR]}
+OUTPUT_BASE = {ENV[CYLC_WORKFLOW_SHARE_DIR]}
+# INPUT_READ_SCRIPT = {ENV[INPUT_READ_SCRIPT]}
+MET_INSTALL_DIR = {ENV[MET_INSTALL_DIR]}
+TIME_START = {ENV[TIME_START]}
+
+[config]
+
+# Documentation for this use case can be found at
+# https://metplus.readthedocs.io/en/latest/generated/met_tool_wrapper/ASCII2NC/ASCII2NC_python_embedding.html
+
+# For additional information, please see the METplus Users Guide.
+# https://metplus.readthedocs.io/en/latest/Users_Guide
+
+###
+# Processes to run
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#process-list
+###
+
+PROCESS_LIST = ASCII2NC
+
+
+###
+# Time Info
+# LOOP_BY options are INIT, VALID, RETRO, and REALTIME
+# If set to INIT or RETRO:
+#   INIT_TIME_FMT, INIT_BEG, INIT_END, and INIT_INCREMENT must also be seti# If set to VALID or REALTIME:
+#   VALID_TIME_FMT, VALID_BEG, VALID_END, and VALID_INCREMENT must also be set
+# LEAD_SEQ is the list of forecast leads to process
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#timing-control
+###
+
+LOOP_BY = INIT
+INIT_TIME_FMT = %Y%m%dT%H
+INIT_BEG = {ENV[TASK_TIME]}
+INIT_END = {ENV[TASK_ENDTIME]}
+INIT_INCREMENT = 6H
+
+LEAD_SEQ = begin_end_incr(0,5,1)
+
+
+###
+# File I/O
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#directory-and-filename-template-info
+###
+
+ASCII2NC_INPUT_DIR =
+ASCII2NC_INPUT_TEMPLATE = "python/{ENV[INPUT_READ_SCRIPT]} {INPUT_BASE}"
+
+ASCII2NC_OUTPUT_DIR =
+ASCII2NC_OUTPUT_TEMPLATE = {OUTPUT_BASE}/obs_{TIME_START}.nc
+
+ASCII2NC_SKIP_IF_OUTPUT_EXISTS = False
+
+ASCII2NC_FILE_WINDOW_BEGIN = 0
+ASCII2NC_FILE_WINDOW_END = 0
+
+
+###########################################
+ASCII2NC_FILE_WINDOW_BEGIN = -1800
+ASCII2NC_FILE_WINDOW_END = 1800
+
+ASCII2NC_INPUT_DIR =
+ASCII2NC_INPUT_TEMPLATE = "python/{ENV[INPUT_READ_SCRIPT]} {INPUT_BASE}/Surface.NiwaList.{init?fmt=%Y%m%dT%2H} {valid?fmt=%Y%m%dT%2H?shift=-10800}00 {ASCII2NC_FILE_WINDOW_BEGIN} {ASCII2NC_FILE_WINDOW_END}"
+
+ASCII2NC_OUTPUT_DIR =
+ASCII2NC_OUTPUT_TEMPLATE = {OUTPUT_BASE}/obs_nc/ascii2nc_python_niwa_{valid?fmt=%Y%m%dT%2H?shift=-10800}00.nc
+
+ASCII2NC_SKIP_IF_OUTPUT_EXISTS = False
+
+
+
+###
+# ASCII2NC Settings
+# https://metplus.readthedocs.io/en/latest/Users_Guide/wrappers.html#ascii2nc
+###
+
+#LOG_ASCII2NC_VERBOSITY = 1
+#ASCII2NC_CONFIG_FILE =
+
+ASCII2NC_WINDOW_BEGIN = 0
+ASCII2NC_WINDOW_END = 0
+
+ASCII2NC_INPUT_FORMAT = python
+
+ASCII2NC_MASK_GRID =
+ASCII2NC_MASK_POLY =
+ASCII2NC_MASK_SID =
+
+ASCII2NC_TIME_SUMMARY_FLAG = False
+ASCII2NC_TIME_SUMMARY_RAW_DATA = False
+ASCII2NC_TIME_SUMMARY_BEG = 000000
+ASCII2NC_TIME_SUMMARY_END = 235959
+ASCII2NC_TIME_SUMMARY_STEP = 300
+ASCII2NC_TIME_SUMMARY_WIDTH = 600
+ASCII2NC_TIME_SUMMARY_GRIB_CODES = 11, 204, 211
+ASCII2NC_TIME_SUMMARY_VAR_NAMES =
+ASCII2NC_TIME_SUMMARY_TYPES = min, max, range, mean, stdev, median, p80
+ASCII2NC_TIME_SUMMARY_VALID_FREQ = 0
+ASCII2NC_TIME_SUMMARY_VALID_THRESH = 0.0
