+++ cset-workflow/app/metplus_ascii2nc/opt/rose-app-niwa.conf
@@ -1,0 +1,9 @@
+[command]
+default=app_env_wrapper run_metplus.py ASCII2NC_python_embedding.conf
+
+[env]
+CONDA_VENV_LOCATION = ${CONDA_METPLUS_VENV_LOCATION}
+METPLUS_OBS_DIR=${METPLUS_OBS_DIR}
+INPUT_READ_SCRIPT=restricted_read_ascii_point_niwa.py
+TASK_TIME=20230705T00
+TASK_ENDTIME=20230706T18
