+++ cset-workflow/app/metplus_ascii2nc/rose-app.conf
@@ -2,7 +2,7 @@
 default=app_env_wrapper run_metplus.py ASCII2NC_python_embedding.conf
 
 [env]
-CONDA_VENV_LOCATION = ${CONDA_METPLUS_VENV_LOCATION}
-METPLUS_OBS_DIR=${METPLUS_OBS_DIR}
+CONDA_VENV_LOCATION=${CONDA_METPLUS_VENV_LOCATION}
 INPUT_READ_SCRIPT=read_ascii_point.py
+METPLUS_OBS_DIR=${METPLUS_OBS_DIR}
 TIME_START=20221004T00
