+++ cset-workflow/app/metplus_point_stat/file/PointStat.conf
@@ -1,0 +1,98 @@
+[dir]
+OUTPUT_BASE = {ENV[CYLC_TASK_WORK_DIR]}
+MET_INSTALL_DIR = {ENV[MET_INSTALL_DIR]}
+TIME_START = {ENV[TIME_START]}
+
+[config]
+
+# Documentation for this use case can be found at
+# https://metplus.readthedocs.io/en/latest/generated/met_tool_wrapper/ASCII2NC/ASCII2NC_python_embedding.html
+
+# For additional information, please see the METplus Users Guide.
+# https://metplus.readthedocs.io/en/latest/Users_Guide
+
+###
+# Processes to run
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#process-list
+###
+
+SCRUB_STAGING_DIR = False
+PROCESS_LIST = PointStat
+LOG_POINT_STAT_VERBOSITY = 4
+
+###
+# Time Info
+# LOOP_BY options are INIT, VALID, RETRO, and REALTIME
+# If set to INIT or RETRO:
+#   INIT_TIME_FMT, INIT_BEG, INIT_END, and INIT_INCREMENT must also be set
+# If set to VALID or REALTIME:
+#   VALID_TIME_FMT, VALID_BEG, VALID_END, and VALID_INCREMENT must also be set
+# LEAD_SEQ is the list of forecast leads to process
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#timing-control
+###
+
+LOOP_BY = INIT
+INIT_TIME_FMT = %Y%m%dT%H
+INIT_BEG = {TIME_START}
+INIT_END = {TIME_START}
+INIT_INCREMENT = 1H
+
+#LEAD_SEQ = 0
+LEAD_SEQ = begin_end_incr(0,6,1)
+
+###
+# File I/O
+# https://metplus.readthedocs.io/en/latest/Users_Guide/systemconfiguration.html#directory-and-filename-template-info
+###
+
+FCST_POINT_STAT_INPUT_DIR = /nesi/project/niwa99999/bornemannjj/CSET_TEST_DATA/model/AKL
+FCST_POINT_STAT_INPUT_TEMPLATE = akl_0p333km_{TIME_START}_met.nc
+
+OBS_POINT_STAT_INPUT_DIR = {ENV[CYLC_WORKFLOW_SHARE_DIR]}/obs_nc
+OBS_POINT_STAT_INPUT_TEMPLATE = ascii2nc_python_niwa_{valid?fmt=%Y%m%d}T{valid?fmt=%2H}00.nc
+
+POINT_STAT_OUTPUT_DIR = {ENV[CYLC_WORKFLOW_SHARE_DIR]}
+POINT_STAT_OUTPUT_TEMPLATE = Point_Stat_{init?fmt=%Y%m%d}T{init?fmt=%2H}00.nc
+
+
+###
+# POINT_STAT Settings
+#https://metplus.readthedocs.io/en/latest/Users_Guide/wrappers.html#pointstat
+###
+
+MODEL = UM
+POINT_STAT_DESC = Tests_NIWA_obs
+
+POINT_STAT_REGRID_TO_GRID = FCST
+POINT_STAT_REGRID_METHOD = NEAREST
+POINT_STAT_REGRID_WIDTH = 1
+POINT_STAT_REGRID_VLD_THRESH = 0.5
+
+POINT_STAT_MASK_GRID = FULL
+
+POINT_STAT_FCST_FILE_TYPE = NETCDF_NCCF
+POINT_STAT_MESSAGE_TYPE = "ADPSFC"
+
+FCST_POINT_STAT_VAR1_NAME = sfc_temp
+FCST_POINT_STAT_VAR1_LEVELS = "({lead?fmt=%H},*,*)"
+FCST_POINT_STAT_VAR1_THRESH = <=273, >273
+#FCST_POINT_STAT_VAR1_OPTIONS = set_attr_name="T2M"; set_attr_level="Z0"
+
+OBS_POINT_STAT_VAR1_NAME = TMP
+OBS_POINT_STAT_VAR1_LEVELS = Z0
+OBS_POINT_STAT_VAR1_THRESH = <=273, >273
+#OBS_POINT_STAT_VAR1_OPTIONS = set_attr_name="T2M"
+
+FCST_POINT_STAT_VAR2_NAME = sfc_rh
+FCST_POINT_STAT_VAR2_LEVELS = "({lead?fmt=%H},*,*)"
+FCST_POINT_STAT_VAR2_THRESH =  <60, >95
+#FCST_POINT_STAT_VAR2_OPTIONS = set_attr_name="RH"; set_attr_level="Z0"
+
+OBS_POINT_STAT_VAR2_NAME = RH
+OBS_POINT_STAT_VAR2_LEVELS = Z0
+OBS_POINT_STAT_VAR2_THRESH = <60, >95
+
+OBS_POINT_STAT_WINDOW_BEGIN = -1800
+OBS_POINT_STAT_WINDOW_END = 1800
+
+POINT_STAT_OUTPUT_FLAG_CNT = BOTH
