+++ cset-workflow/app/metplus_point_stat/opt/rose-app-niwa.conf
@@ -1,0 +1,7 @@
+[command]
+default=app_env_wrapper run_metplus.py restricted_PointStat_cycling_niwa.conf restricted_PointStat_science_niwa.conf
+
+[env]
+CONDA_VENV_LOCATION = ${CONDA_METPLUS_VENV_LOCATION}
+METPLUS_FCST_DIR = ${METPLUS_FCST_DIR}
+TIME_START=20230705T00
