+++ cset-workflow/app/metplus_point_stat/rose-app.conf
@@ -2,6 +2,6 @@
 default=app_env_wrapper run_metplus.py PointStat.conf
 
 [env]
-CONDA_VENV_LOCATION = ${CONDA_METPLUS_VENV_LOCATION}
+CONDA_VENV_LOCATION=${CONDA_METPLUS_VENV_LOCATION}
 METPLUS_FCST_DIR=${METPLUS_FCST_DIR}
 TIME_START=20221004T00
