--- cset-workflow/app/populate_website/rose-app.conf
@@ -1,2 +1,0 @@
-[command]
-default=app_env_wrapper populate-website.py
