+++ cset-workflow/app/run_cset_recipe/bin/run-cset-recipe.py
@@ -2,23 +2,57 @@
 
 """Run a recipe with the CSET CLI."""
 
+import fcntl
 import json
 import logging
 import os
 import subprocess
+import zipfile
 from pathlib import Path
 from uuid import uuid4
 
-import redis
-from ruamel.yaml import YAML
-
 logging.basicConfig(level=logging.INFO)
 
-r = redis.Redis().from_url(
-    Path(f"{os.getenv('CYLC_WORKFLOW_SHARE_DIR')}/redis_uri")
-    .read_text(encoding="utf-8")
-    .strip()
-)
+
+def combine_dicts(d1: dict, d2: dict) -> dict:
+    """Recursively combines two dictionaries.
+
+    Duplicate atoms favour the second dictionary.
+    """
+    # TODO: Test this function, as I'm not sure it is 100% correct.
+    # Update existing keys.
+    for key in d1.keys() & d2.keys():
+        if isinstance(d1[key], dict):
+            d1[key] = combine_dicts(d1[key], d2[key])
+        else:
+            d1[key] = d2[key]
+    # Add any new keys.
+    for key in d2.keys() - d1.keys():
+        d1[key] = d2[key]
+    return d1
+
+
+def append_to_index(index_path: Path, record: dict):
+    """Append the plot record to the index file.
+
+    Record should have the form {"Model Name": {"Plot Name": "directory-uuid"}}
+    """
+    with open(index_path, "a+t", encoding="UTF-8") as fp:
+        # Lock file until closed.
+        fcntl.flock(fp, fcntl.LOCK_EX)
+        # Open in append mode then seek back to avoid errors if the file does
+        # not exist.
+        fp.seek(0)
+        try:
+            index = json.load(fp)
+        except json.JSONDecodeError:
+            index = {}
+        index = combine_dicts(index, record)
+        fp.seek(0)
+        fp.truncate()
+        json.dump(index, fp)
+
+
 input_file = (
     Path(
         f"{os.getenv('CYLC_WORKFLOW_SHARE_DIR')}/cycle/{os.getenv('CYLC_TASK_CYCLE_POINT')}/input_path"
@@ -26,30 +60,29 @@ input_file = (
     .read_text(encoding="utf-8")
     .strip()
 )
-output_file = f"{os.getcwd()}/{uuid4()}"
+plot_id = str(uuid4())
+output_directory = Path.cwd() / plot_id
 
 # Takes recipe from CSET_RECIPE environment variable.
-subprocess.run(("cset", "-v", "bake", input_file, output_file), check=True)
+subprocess.run(("cset", "-v", "bake", input_file, output_directory), check=True)
 
-with YAML(typ="safe", pure=True) as yaml:
-    recipe = yaml.load(os.getenv("CSET_RECIPE"))
+with open(output_directory / "meta.json", "rt", encoding="UTF=8") as fp:
+    recipe_meta = json.load(fp)
 
-# TODO: Get actual model here.
-source_model = "Model #1"
+title = recipe_meta.get("title", "Unknown")
+# TODO: Save model in meta.json. Currently this always returns "Unknown".
+source_model = recipe_meta.get("model", "Unknown")
 
-# TODO: Create actual diagnostic log file.
-Path(output_file + ".log").write_text("Not yet implemented.")
+# TODO: Give archive a meaningful name, probably a slugified title.
+archive_path = output_directory / "diagnostic.zip"
+with zipfile.ZipFile(archive_path, "w", compression=zipfile.ZIP_DEFLATED) as archive:
+    for path in output_directory.rglob("*"):
+        # Check the archive doesn't add itself.
+        if not path.samefile(archive_path):
+            archive.write(path, arcname=path.relative_to(output_directory))
 
-# Write plot record json
-plot_record = json.dumps(
-    {
-        "model": source_model,
-        "title": recipe["title"],
-        "description": recipe["description"],
-        "plot_data_location": output_file + ".nc",
-        "plot_location": output_file + ".svg",
-        "raw_data_path": input_file,
-        "diagnostic_log_location": output_file + ".log",
-    }
-)
-r.lpush("plot_records", plot_record)
+# Symbolic link to output from plots directory.
+webdir_path = Path(f"{os.getenv('WEB_DIR')}/plots/{plot_id}")
+webdir_path.symlink_to(output_directory, target_is_directory=True)
+# TODO: Consider making plot_id the key to handle duplicated titles.
+append_to_index(webdir_path.parent / "index.json", {source_model: {title: plot_id}})
