+++ cset-workflow/flow.cylc
@@ -14,22 +14,21 @@ URL = https://metoffice.github.io/CSET-workflow
     # Only runs on the first cycle.
     R1/^ = """
     build_conda => install_website_skeleton
-    build_conda => install_local_cset => db
+    build_conda => install_local_cset
     """
 
     # Runs every cycle. Include files get inserted into this graph.
     P1 = """
-    db[1]:ready => fetch_fcst => recipe_start
+    install_local_cset[1] => fetch_fcst
+    install_website_skeleton[1] => recipe_start
+    fetch_fcst => recipe_start
     recipe_start => recipe_finish
-    recipe_finish => populate_website => housekeeping
-    install_website_skeleton[1] => populate_website
-    populate_website[-P1] => populate_website
+    recipe_finish => housekeeping
     """
 
     # Only runs on the final cycle.
     R1/$ = """
-    populate_website => shutdown_db => send_email
-    populate_website => finish_website
+    recipe_finish => finish_website => send_email
     """
 
 
@@ -38,20 +37,18 @@ URL = https://metoffice.github.io/CSET-workflow
     script = rose task-run -v
     execution time limit = PT15M
         [[[environment]]]
-        # Load modules/conda environment for dependencies.
         CSET_ENV_USE_MODULES = {{CSET_ENV_USE_MODULES}}
         {% if CSET_ENV_USE_MODULES %}
         MODULES_LIST = {{MODULES_LIST}}
         MODULES_PURGE = {{MODULES_PURGE}}
         {% endif %}
+
         CSET_ENV_USE_CONDA = {{CSET_ENV_USE_CONDA}}
         {% if CSET_ENV_USE_CONDA %}
+        CONDA_PATH = {{CONDA_PATH}}
         CONDA_VENV_LOCATION = {{CONDA_VENV_LOCATION}}
         {% endif %}
-        CSET_ENV_USE_LOCAL_CSET = {{CSET_ENV_USE_LOCAL_CSET}}
-        {% if CSET_ENV_USE_LOCAL_CSET %}
-        CSET_LOCAL_CSET_PATH = {{CSET_LOCAL_CSET_PATH}}
-        {% endif %}
+
         CSET_ENV_SEPARATE_MET = {{CSET_ENV_SEPARATE_MET}}
         {% if CSET_ENV_SEPARATE_MET %}
         CONDA_METPLUS_VENV_LOCATION = {{CONDA_METPLUS_VENV_LOCATION}}
@@ -61,49 +58,36 @@ URL = https://metoffice.github.io/CSET-workflow
         MET_LIBRARIES = {{MET_LIBRARIES}}
         {% endif %}
 
+        WEB_DIR = {{WEB_DIR}}
+
     [[build_conda]]
     execution time limit = PT30M
         [[[environment]]]
         CONDA_VENV_CREATE = {{CONDA_VENV_CREATE}}
-        CONDA_PATH = {{CONDA_PATH}}
 
     [[install_local_cset]]
     execution time limit = PT5M
         [[[environment]]]
         CSET_ENV_USE_LOCAL_CSET = {{CSET_ENV_USE_LOCAL_CSET}}
-
-    [[db]]
-    {% if DB_LONG_JOB %}
-    execution time limit = PT3D
-    {% else %}
-    execution time limit = PT6H
-    {% endif %}
-        [[[outputs]]]
-        ready = "Database ready"
+        {% if CSET_ENV_USE_LOCAL_CSET %}
+        CSET_LOCAL_CSET_PATH = {{CSET_LOCAL_CSET_PATH}}
+        {% endif %}
 
     [[install_website_skeleton]]
     # Copies the static files that make up the web interface.
         [[[environment]]]
-        WEB_DIR = {{WEB_DIR}}
         CLEAN_WEB_DIR = {{CLEAN_WEB_DIR}}
 
     [[finish_website]]
     # Updates the workflow info in the web interface.
     platform = localhost
         [[[environment]]]
-        WEB_DIR = {{WEB_DIR}}
 
     [[fetch_fcst]]
     # Fetch data from disk or a file based archival system.
         [[[environment]]]
         ROSE_APP_OPT_CONF_KEYS = {{FETCH_FCST_OPT_CONF}}
         FILE_PATH = {{CSET_INPUT_FILE_PATH}}
-
-    [[populate_website]]
-    # Fill the web interface with content from the DB.
-    execution time limit = PT5M
-        [[[environment]]]
-        WEB_DIR = {{WEB_DIR}}
 
     [[recipe_start]]
     platform = localhost
@@ -113,21 +97,11 @@ URL = https://metoffice.github.io/CSET-workflow
     platform = localhost
     script = true
 
-    [[shutdown_db]]
-    platform = localhost
-
     [[send_email]]
     # Send email to notify that the workflow is complete.
     platform = localhost
         [[[environment]]]
         WEB_ADDR = {{WEB_ADDR}}
-        WEB_DIR = {{WEB_DIR}}
-
-    [[clean_web_dir]]
-    # Removes any files in the web dir.
-    platform = localhost
-        [[[environment]]]
-        WEB_DIR = {{WEB_DIR}}
 
     [[housekeeping]]
         [[[environment]]]
@@ -141,8 +115,17 @@ URL = https://metoffice.github.io/CSET-workflow
 {% if DETERMINISTIC_PLOT_SURFACE_AIR_TEMP %}
 {% include 'includes/deterministic_plot_surface_air_temp.cylc' %}
 {% endif %}
+{% if DETERMINISTIC_PLOT_AGGREGATE_PRECIPITATION %}
+{% include 'includes/deterministic_plot_aggregate_precipitation.cylc' %}
+{% endif %}
+{% if DETERMINISTIC_PLOT_MODEL_LEVEL_AIR_TEMP %}
+{% include 'includes/deterministic_plot_model_level_air_temp.cylc' %}
+{% endif %}
 {% if ENSEMBLE_PLOT_SURFACE_AIR_TEMP %}
 {% include 'includes/ensemble_plot_surface_air_temp.cylc' %}
+{% endif %}
+{% if MEAN_HOURLY_ACCUMULATED_PRECIPITATION_ENSEMBLE_POSTAGE_STAMP_PLOT %}
+{% include 'includes/mean_hourly_accumulated_precip_ensemble_postage_stamp_plot.cylc' %}
 {% endif %}
 
 {# Site-specific details that add to (or override) the core suite definition #}
