+++ cset-workflow/includes/deterministic_plot_aggregate_precipitation.cylc
@@ -1,46 +1,43 @@
 [scheduling]
     [[graph]]
     P1 = """
-    recipe_start => deterministic_plot_surface_air_temp => recipe_finish
+    recipe_start => deterministic_plot_aggregate_precipitation => recipe_finish
     """
 
 [runtime]
-    [[deterministic_plot_surface_air_temp]]
+    [[deterministic_plot_aggregate_precipitation]]
     script = rose task-run -v --app-key=run_cset_recipe
     execution time limit = PT15M
         [[[environment]]]
         CSET_RECIPE = """
-title: Plot Instant Air Temperature
+title: Extract sub-hourly or hourly ppn and aggregate to user defined intervals
 description: |
-  <p>Extracts and plots the instantaneous 1.5m air temperature from a file.</p>
-  <p>The time coordinate is meaned.</p>
-  <p>These descriptions are HTML, so you can do things like <b>bold</b>,
-  <i>italics</i>, <span style=\"color: red; font-family: cursive;\">and more</span>.
+  <p>Extracts out the sub-hourly or hourly stratiform rainfall rate from a file and aggregates it up to new intervals and writes it to a new one.</p>
+  <p>The time coordinate is accumulated.</p>
   </p>
 
 steps:
   - operator: read.read_cubes
-    constraint:
-      operator: constraints.generate_stash_constraint
-      stash: m01s03i236
 
   - operator: filters.filter_cubes
     constraint:
       operator: constraints.combine_constraints
-      constraint:
+      stash_constraint:
         operator: constraints.generate_stash_constraint
-        stash: m01s03i236
+        stash: m01s04i203
       cell_methods_constraint:
         operator: constraints.generate_cell_methods_constraint
         cell_methods: []
 
-  - operator: collapse.collapse
+  - operator: aggregate.time_aggregate
     coordinate: time
-    method: MEAN
+    method: SUM
+    interval_iso: PT2H
 
   - operator: plot.spatial_contour_plot
     file_path: CSET_OUTPUT_PATH
 
   - operator: write.write_cube_to_nc
+    # This is a magic value that becomes the runtime output file path.
     file_path: CSET_OUTPUT_PATH
 """
