+++ cset-workflow/meta/rose-meta.conf
@@ -73,7 +73,7 @@ sort-key=env-032
 ns=Environment
 description=Location of the conda virtual environment.
 help=The can be found by running `conda env list` and copying the path displayed
-    there.
+    there. If unspecified it defaults to the workflow share directory.
 type=quoted
 sort-key=env-033
 
@@ -236,6 +236,22 @@ type=python_boolean
 compulsory=true
 sort-key=diag-040
 
+[template variables=DETERMINISTIC_PLOT_AGGREGATE_PRECIPITATION]
+ns=Diagnostics
+description=Create a plot for time aggregated precipitation.
+help=See includes/deterministic_plot_aggregate_precipitation.cylc
+type=python_boolean
+compulsory=true
+sort-key=diag-040
+
+[template variables=DETERMINISTIC_PLOT_MODEL_LEVEL_AIR_TEMP]
+ns=Diagnostics
+description=Create a plot for time aggregated air temperature on model level.
+help=See includes/deterministic_plot_model_level_air_temp.cylc
+type=python_boolean
+compulsory=true
+sort-key=diag-040
+
 [template variables=ENSEMBLE_PLOT_SURFACE_AIR_TEMP]
 ns=Diagnostics
 description=Create postage-stamp plots for the time meaned surface air temperature for an ensemble forecast.
@@ -243,3 +259,11 @@ help=See includes/ensemble_plot_surface_air_temp.cylc
 type=python_boolean
 compulsory=true
 sort-key=diag-050
+
+[template variables=MEAN_HOURLY_ACCUMULATED_PRECIPITATION_ENSEMBLE_POSTAGE_STAMP_PLOT]
+ns=Diagnostics
+description=Create postage-stamp plots for the hourly accumulated precipitation.
+help=See includes/mean_hourly_accumulated_precip_ensemble_postage_stamp_plot.cylc
+type=python_boolean
+compulsory=true
+sort-key=diag-060
