+++ cset-workflow/requirements/environment.yml
@@ -3,11 +3,8 @@ channels:
   - conda-forge
 dependencies:
   # Runtime dependencies
-  - redis-server > 6.2
-  - redis-py >= 4.1.0
-  - cset >= 0.3.0
+  - cset >= 0.4.0
   - pip >= 23.1
-  - mo_pack >= 0.2.0
 
   # Development dependencies
   - pre-commit
