+++ cset-workflow/rose-suite.conf.example
@@ -11,11 +11,13 @@ CSET_ENV_USE_MODULES=False
 CSET_INPUT_FILE_PATH=""
 !!CSET_LOCAL_CSET_PATH=""
 DB_LONG_JOB=False
+DETERMINISTIC_PLOT_AGGREGATE_PRECIPITATION=False
+DETERMINISTIC_PLOT_MODEL_LEVEL_AIR_TEMP=False
 DETERMINISTIC_PLOT_SURFACE_AIR_TEMP=False
 ENSEMBLE_PLOT_SURFACE_AIR_TEMP=False
 FETCH_FCST_OPT_CONF="filesystem"
-HOUSEKEEPING_MODE=1
-!!CONDA_METPLUS_VENV_LOCATION=""
+HOUSEKEEPING_MODE=0
+MEAN_HOURLY_ACCUMULATED_PRECIPITATION_ENSEMBLE_POSTAGE_STAMP_PLOT=False
 !!METPLUS_BASE=""
 !!METPLUS_OBS_DIR=""
 !!METPLUS_OPT_CONFIG_KEYS=""
@@ -24,6 +26,6 @@ METPLUS_POINT_STAT=False
 !!MET_LIBRARIES=""
 !!MODULES_LIST=
 !!MODULES_PURGE=True
-WEB_ADDR=""
-WEB_DIR=""
 SITE=""
+WEB_ADDR=""
+WEB_DIR="$HOME/public_html/CSET"
